+++ config/entries.js
@@ -1,0 +1,4 @@
+module.exports = {
+  app: ['./src/js/index.js', './src/scss/style.scss'],
+  editor: ['./src/js/editor.js', './src/scss/editor.scss'],
+}